Refactor: Optimize EmbeddedResource and DTO performance
EmbeddedResource Enhancements:
- Add directory caching to reduce filesystem operations
- Optimize decompression with buffer preallocation
- Improve path caching with base directory hints
- Add performance monitoring with timing and memory logs
- Fix thread-safety issues in caching mechanisms

DTO Performance Optimizations:
- Implement string interning for memory optimization
- Add batch processing for large collections
- Introduce type aliases in CodebookDto and GmodDto for readability
- Update JSON deserialization to return std::optional for error handling
- Enhance immutability by deleting copy/move assignment operators
- Add safe array handling for RapidJSON serialization

Testing Infrastructure:
- Standardize test file naming (remove _test suffix)
- Consolidate test organization and structure
- Add LocationsTests for comprehensive coverage

+++ cpp/include/dnv/vista/sdk/Codebooks.h
@@ -66,7 +66,7 @@ namespace dnv::vista::sdk
 			 * @param codebooks Pointer to the codebooks array
 			 * @param index Starting position index
 			 */
-			Iterator( const std::array<Codebook, static_cast<size_t>( NUM_CODEBOOKS )>* codebooks, size_t index );
+			Iterator( const std::array<Codebook, NUM_CODEBOOKS>* codebooks, size_t index );
 
 			/**
 			 * @brief Dereference operator
@@ -140,7 +140,6 @@ namespace dnv::vista::sdk
 		//-------------------------------------------------------------------
 		// Construction / Destruction
 		//-------------------------------------------------------------------
-
 		/** @brief Default constructor */
 		Codebooks() = default;
 
@@ -247,6 +246,6 @@ namespace dnv::vista::sdk
 		VisVersion m_visVersion;
 
 		/** @brief Array of codebooks */
-		std::array<Codebook, static_cast<size_t>( NUM_CODEBOOKS )> m_codebooks;
+		std::array<Codebook, NUM_CODEBOOKS> m_codebooks;
 	};
 }
